styles the carousel fixes typos
