+++ cypress/fixtures/about_us_view_sections.json
@@ -36,7 +36,7 @@
     },
     {
       "id": 4,
-      "variant": "carusel",
+      "variant": "carousel",
       "header": "Our Partners",
       "cards": [
         {
